server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
google-generativeai==0.3.2
Pillow==10.1.0
python-dotenv==1.0.0
pytesseract==0.3.10

